Server/requirements.txt
pyjwt
requests
jsonpickle
Flask-SQLAlchemy
pymongo
pika
flask
apscheduler
mysql-connector-python
flask-cors